refactor: always register with license key, if available
